fixed yet another printf() format warning
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@20179 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/common/docview.cpp
@@ -2115,7 +2115,7 @@ void wxFileHistory::Save(wxConfigBase& config)
     for (i = 0; i < m_fileHistoryN; i++)
     {
         wxString buf;
-        buf.Printf(wxT("file%d"), i+1);
+        buf.Printf(wxT("file%d"), (int)i+1);
         config.Write(buf, wxString(m_fileHistory[i]));
     }
 }
